+++ Assignment3/spl-net/src/main/java/bgu/spl/net/api/Messages/ClientMessages/KDAMCHECK.java
@@ -6,9 +6,9 @@ import bgu.spl.net.api.Messages.ServerMessages.Error;
 import bgu.spl.net.api.Users.User;
 
 public class KDAMCHECK extends Message {
-    private int courseNum;
+    private short courseNum;
 
-    public KDAMCHECK(int courseNum){
+    public KDAMCHECK(short courseNum){
         super(Short.parseShort("6"));
         this.courseNum = courseNum;
     }
